fourth ex is complete

+++ tasks/task1/task1/Program.cs
@@ -1,6 +1,9 @@
 ﻿using System;
+using System.Collections;
+using System.Collections.Generic;
 using System.IO;
 using System.Linq;
+using System.Threading;
 
 namespace task1
 {
@@ -10,7 +13,7 @@ namespace task1
         {
             var path = Console.ReadLine();
             const string backUp = "backup.txt";
-            if (path != null)
+            if (path != null && File.Exists(path))
             {
                 File.Copy(path, backUp);
                 var fileIn = new StreamReader(path);
@@ -19,7 +22,6 @@ namespace task1
                 if (word != null && fileText.Contains(word))
                 {
                     var substring = fileText.Replace(word, "");
-                    Console.WriteLine(substring);
                     fileIn.Close();
                     using (var sw = new StreamWriter(path, false))
                     {
@@ -36,7 +38,7 @@ namespace task1
         private static void Second()
         {
             var path = Console.ReadLine();
-            if (path != null)
+            if (path != null && File.Exists(path))
             {
                 var fileIn = new StreamReader(path);
                 var pattern = new[] {',', '-', '.', '"', ':', '\t', '!', '?', ' ', '(', ')'};
@@ -62,7 +64,7 @@ namespace task1
         private static void Third()
         {
             var path = Console.ReadLine();
-            if (path != null)
+            if (path != null && File.Exists(path))
             {
                 var fileIn = new StreamReader(path);
                 var text = fileIn.ReadToEnd().Split('.');
@@ -76,12 +78,67 @@ namespace task1
                 Console.WriteLine("Something wrong with the path");
         }
 
-    public static void Main(string[] args)
+        private static void Fourth()
         {
-            //First();
-            //File.Delete("backUp.txt");
-            //Second();
-            Third();
+            var path = Console.ReadLine();
+            while (true)
+            {
+                var catalog = new ArrayList();
+                if (path != null && Directory.Exists(path))
+                {
+                    var directories = Directory.GetDirectories(path);
+                    var files = Directory.GetFiles(path);
+
+                    foreach (var d in directories) catalog.Add(d);
+                    foreach (var f in files) catalog.Add(f);
+                }
+
+                catalog.Sort();
+                for (var i = 0; i < catalog.Count; i++)
+                {
+                    if (Directory.Exists(catalog[i].ToString())) Console.Write(i + " - ");
+                    Console.WriteLine("\t" + catalog[i]);
+                }
+
+                string pointer;
+                do
+                {
+                    Console.WriteLine("Enter index of the next directories or 'exit' to quit program:");
+                    pointer = Console.ReadLine();
+                    try
+                    {
+                        var flag = !Directory.Exists(catalog[Convert.ToInt32(pointer)].ToString());
+                    }
+                    catch
+                    {
+                        Console.WriteLine("U choose wrong index");
+                    }
+                    if (pointer == "exit") return;
+                } while (Convert.ToInt32(pointer) > catalog.Count || Convert.ToInt32(pointer) < 0);
+                path = catalog[Convert.ToInt32(pointer)].ToString();
+                Console.Clear();
+            }
+        }
+
+        public static void Main(string[] args)
+        {
+            Console.WriteLine("To run first method which delete one word from file enter 1\nTo run second method that will show u count of a word in file and also show u every 10th word enter 2\nTo run third method which will show u third sentence and reverse all character in word reverse enter 3\nTo run Fourth method which open u catalog by address that u type enter \n");
+            var flag = Console.ReadLine();
+            switch (flag)
+            {
+                case "1": First();
+                    File.Delete("backup.txt");
+                    break;
+                case "2": Second();
+                    break;
+                case "3": Third();
+                    break;
+                case "4": Fourth();
+                    break;
+                default:
+                    Console.WriteLine("U write wrong number of function");
+                    break;
+            }
         }
     }
 }
